wip: migrate session leaderboard to new architecture
