+++ apps/frontend-pwa/src/components/common/LiveQuizLeaderboard.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@apollo/client'
 import {
-  GetSessionLeaderboardDocument,
+  GetLiveQuizLeaderboardDocument,
   SelfDocument,
 } from '@klicker-uzh/graphql/dist/ops'
 import Loader from '@klicker-uzh/shared-components/src/Loader'
@@ -15,28 +15,26 @@ import Rank1Img from '../../../public/rank1.svg'
 import Rank2Img from '../../../public/rank2.svg'
 import Rank3Img from '../../../public/rank3.svg'
 
-interface SessionLeaderboardProps {
-  sessionId: string
-  className?: string
-}
-
 type BlockResult = {
   score: number
   rank: number
 } | null
 
-function SessionLeaderboard({
-  sessionId,
+function LiveQuizLeaderboard({
+  quizId,
   className,
-}: SessionLeaderboardProps): React.ReactElement {
+}: {
+  quizId: string
+  className?: string
+}): React.ReactElement {
   const t = useTranslations()
 
   const { data: selfData } = useQuery(SelfDocument, {
     fetchPolicy: 'cache-only',
   })
 
-  const { data, loading } = useQuery(GetSessionLeaderboardDocument, {
-    variables: { sessionId },
+  const { data, loading } = useQuery(GetLiveQuizLeaderboardDocument, {
+    variables: { quizId },
     // use network-only to trigger a refetch once the component is displayed
     // TODO: replace this by a send of the leaderboard within the subscription
     // TODO: otherwise, this could overload the server if 1000 simultaneous users are on the leaderboard
@@ -48,7 +46,7 @@ function SessionLeaderboard({
   // save the current leaderboard to local storage
   useEffect(() => {
     const asyncFunc = async () => {
-      const leaderboard = data?.sessionLeaderboard
+      const leaderboard = data?.liveQuizLeaderboard
 
       const selfEntry = leaderboard?.find(
         (entry) => entry.participantId === selfData?.self?.id
@@ -87,14 +85,14 @@ function SessionLeaderboard({
     return <Loader />
   }
 
+  const leaderboard = data.liveQuizLeaderboard ?? []
   return (
     <div className={twMerge('space-y-4', className)}>
       <H2>{t('shared.leaderboard.lqLeaderboard')}</H2>
       <div>
-        {data.sessionLeaderboard?.length &&
-        data.sessionLeaderboard.length > 0 ? (
+        {leaderboard.length && leaderboard.length > 0 ? (
           <Leaderboard
-            leaderboard={data.sessionLeaderboard ?? []}
+            leaderboard={leaderboard ?? []}
             participant={selfData?.self}
             podiumImgSrc={{
               rank1: Rank1Img,
@@ -142,4 +140,4 @@ function SessionLeaderboard({
   )
 }
 
-export default SessionLeaderboard
+export default LiveQuizLeaderboard
