+++ packages/graphql/src/public/client.json
@@ -128,6 +128,7 @@
    "GetGroupActivitySummary": "7a3935db0e7cae8d5baf574a1c0506f568ffaaae1522a49363245946cff64d2d",
    "GetLecturerViewLiveQuiz": "d1ab35eb40e5193d9ef4a4563d1057e860db1b9774d82076ccc1c380350139ff",
    "GetLiveQuizHMAC": "f172839e4e15bdd42039bd90c6f5ecace80f0d2ce00f34bedaceed9a4dbe39d4",
+   "GetLiveQuizLeaderboard": "7f849e9a714a388215d6c7cce9582debe85faa83b11871b330de7261aaf6205d",
    "GetLiveQuizSummary": "fd17c092be3d096abd6d31e89d8249bd4b482a2a6310f6893e9e32c5a70666c0",
    "GetLoginToken": "491946fc1fcbd066ea481c564313f4e50b7cc98ea16ab16a7a2146d75ec8cf00",
    "GetMicroLearningEvaluation": "6c68d74ead2c930449faab6b9e7979b12b00b90330aa5c3c7c7898289b47cc59",
@@ -145,8 +146,7 @@
    "GetPublicParticipantProfile": "840169ba8a67902da899e1f08c92668228657ba5c6af024e42737081179bb343",
    "GetRunningLiveQuiz": "07b3907d0c5847836feeaded8603b142dc66bbb5076546005e584c38bdfaeaf9",
    "GetRunningSessions": "db9601297d4de1b497598ff8a2a58c286492ec65cec0116b63f630d1bd8c86b7",
-   "GetSessionEvaluation": "3781c9e476d64a0f39d7222a509eaf7e62e237a7a8093519262a3eff2c9204ec",
-   "GetSessionLeaderboard": "09404af10f91c8da107743b4ba18a43e35507163176a8cd8ac6542feaa08bb91",
+   "GetSessionEvaluation": "a5ad7994bf1d2baaf82c03582690db52cf8daefe3d23b5b3953e54b20db4d48c",
    "GetSingleCourse": "cbbfd88b1c8e1f9c0200f568c0639105f85d7c6bc8c23a0af0920f1b2a2230c8",
    "GetSingleLiveQuiz": "4e722dcf7beaf901f85ef30e39263e2432bef21fbb44bf0a37426c04321d512d",
    "GetSingleMicroLearning": "c24a6c26ed596ad1aabec4f6a1daef531aa57f769aa875875b2d126d0e73f64c",
